feat(discord): Login by discord token and new application id
An option to log in to Discord using a token has been added. This is an advanced login method that speeds up the process.

The dialog appears identical to the one for logging in with a YT Music token, and this option disappears after logging in.

The app ID has also been changed to the dedicated Discord app.
(The Metrolist logo will now be displayed if there's an error obtaining a song cover art on Discord.)

Closes #1247